select changes and bug fixes

+++ src/input.js
@@ -2,8 +2,12 @@ export default class Input{
     constructor(command){
         this.moveState = 0;//track when type, 0 = none, 1 = field move, 2 = object move, 3 = object position set, 4 = vector change, 5 = timeline change
         var adding = false;//track when add button is pressed
+        this.shifting = false;
         //
         this.overState = 0;//track what the mouse is over, 0 = field, 1 = header, 2 = left column, 3= timeline
+        //
+        this.propsState = false;//is a props field selected
+        this.propActive = null;
         //
         this.command = command;
         this.timeline = command.timeline;
@@ -41,10 +45,19 @@ export default class Input{
                     }
                     break;
                 case "ArrowRight":
-                    command.retime(.1);
+                    if(!this.propsState){
+                        command.retime(.1);
+                    }
                     break;
                 case "ArrowLeft":
-                    command.retime(-.1);
+                    if(!this.propsState){
+                        command.retime(-.1);
+                    }
+                    break;
+                case "ArrowUp":
+                    if(this.propsState){
+                        event.preventDefault();
+                    }
                     break;
                 case " ":
                     //console.log("Not working");
@@ -52,10 +65,21 @@ export default class Input{
                     window.requestAnimationFrame(command.loopStart);
                     break;
                 case "Backspace":
-                    input.selected.delete();
+                    if(!this.propsState){
+                        input.selected.delete();
+                    }
                     break;
                 case "l":
                     input.selected.toggleLock();
+                    break;
+                case "Enter":
+                    if(this.propsState){
+                        this.propActive.blur();
+                        this.propsState = false;
+                    }
+                    break;
+                case "Shift":
+                    this.shifting = true;
                     break;
             }
         });
@@ -77,12 +101,16 @@ export default class Input{
                     break;
                 case "Enter":
                     break;
+                case "Shift":
+                    this.shifting = false;
+                    break;
             }
         })
         //
         command.svg.on("mousedown", function(){
             let mouse = d3.mouse(this);
             console.log(input.moveState);
+            input.propsState = false;
             //
             if(input.moveState == 0){
                 input.moveState = 1;
@@ -103,7 +131,7 @@ export default class Input{
                     command.repos(emx - mX, emy - mY);
                     break;
                 case 2:
-                    this.active.repos(emx, emy);
+                    this.command.reposObj(emx, emy);
                     break;
                 case 3: 
                     this.active.repos(emx, emy);
@@ -127,12 +155,19 @@ export default class Input{
         //
         document.addEventListener("mouseup", event => {//this is kinda broken
             if(this.moveState == 3){
-                command.updateVectors(input.active, 1);
+                command.vectorMode = 1;
+                command.updateVectors();
                 this.moveState = 0;
                 //input.velConf = true;
-                this.active.nets[0].self.show();
+                //this.active.nets[0].self.show();
                 command.update();
             }else{
+                console.log(`mx: ${mX}, stX: ${stX}`);
+                if((mX - Math.ceil(stX) == 0) && (mY - Math.ceil(stY) == 0) && input.moveState != 2){//click, no movement (y start has to be rounded for some reason)
+                    console.log("trying to hide");
+                    command.select();
+                    this.active = null;
+                }
                 this.moveState = 0;
             }
         });
@@ -141,7 +176,8 @@ export default class Input{
             input.moveState = 5;
             command.time = command.timeline.timeX.invert(d3.mouse(this)[0]);
             command.retime(0);
-            //console.log("huh");
+            input.propsState = false;
+            console.log("huh");
         });
         //
     }
@@ -156,6 +192,14 @@ export default class Input{
             input.active = obj;
             input.selected = obj;
             obj.command.retime(0);
+            if(input.shifting){
+                console.log("Shifting");
+                input.command.shiftSelect(obj);
+                obj.self.raise();
+            }else{
+                input.command.select(obj);
+                obj.self.raise();
+            }
         });
     }
     //
@@ -168,4 +212,86 @@ export default class Input{
             }
         });
     }
+    //
+    props(command, self){
+        var input = this;
+        self.posx.on("click", function(){
+            this.select();
+            input.propsState = true;
+            input.propActive = this;
+        });
+        self.posy.on("click", function(){
+            this.select();
+            input.propsState = true;
+            input.propActive = this;
+        });
+        self.velx.on("click", function(){
+            this.select();
+            input.propsState = true;
+            input.propActive = this;
+        });
+        self.vely.on("click", function(){
+            this.select();
+            input.propsState = true;
+            input.propActive = this;
+        });
+        self.accelx.on("click", function(){
+            this.select();
+            input.propsState = true;
+            input.propActive = this;
+        });
+        self.accely.on("click", function(){
+            this.select();
+            input.propsState = true;
+            input.propActive = this;
+        });
+        //
+        self.posx.on("input", function(){
+            if(isNumeric(this.value)){
+                command.selected.profile.setValues(0, parseFloat(this.value), command.selected.py);
+                command.objUpdate(command.selected, true);
+            }
+        });
+        //
+        self.posy.on("input", function(){
+            if(isNumeric(this.value)){
+                command.selected.profile.setValues(0, command.selected.px, parseFloat(this.value));
+                command.objUpdate(command.selected, true);
+            }
+        });
+        //
+        self.velx.on("input", function(){
+            if(isNumeric(this.value)){
+                command.selected.profile.setValues(1, parseFloat(this.value), parseFloat(self.vely.property("value")));
+                command.objUpdate(command.selected, true);
+            }
+        });
+        //
+        self.vely.on("input", function(){
+            if(isNumeric(this.value)){
+                command.selected.profile.setValues(1, parseFloat(self.velx.property("value")), parseFloat(this.value));
+                command.objUpdate(command.selected, true);
+            }
+        });
+        //
+        self.accelx.on("input", function(){
+            if(isNumeric(this.value)){
+                command.selected.profile.setValues(2, parseFloat(this.value), parseFloat(self.accely.property("value")));
+                command.objUpdate(command.selected, true);
+            }
+        });
+        //
+        self.accely.on("input", function(){
+            if(isNumeric(this.value)){
+                command.selected.profile.setValues(2, parseFloat(self.accelx.property("value")), parseFloat(this.value));
+                command.objUpdate(command.selected, true);
+            }
+        });
+    }
+}
+
+function isNumeric(str) {
+    if (typeof str != "string") return false // we only process strings!  
+    return !isNaN(str) && // use type coercion to parse the _entirety_ of the string (`parseFloat` alone does not do this)...
+           !isNaN(parseFloat(str)) // ...and ensure strings of whitespace fail
 }
